+++ CHANGELOG.md
@@ -20,6 +20,8 @@ Main (unreleased)
 
 - Add `otelcol.exporter.googlecloudpubsub` community component to export metrics, traces, and logs to Google Cloud Pub/Sub topic. (@eraac)
 
+- Send remote config status to the remote server for the remotecfg service. (@erikbaranowski)
+
 ### Enhancements
 
 - Add support of `tls` in components `loki.source.(awsfirehose|gcplog|heroku|api)` and `prometheus.receive_http` and `pyroscope.receive_http`. (@fgouteroux)